% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{check_master_index_cols}
\alias{check_master_index_cols}
\title{Check Master Index Columns}
\usage{
check_master_index_cols(.tab)
}
\arguments{
\item{.tab}{Tibble to validate}
}
\value{
Logical indicating if columns are correct
}
\description{
Validates the column names of a master index data frame.
}
\keyword{internal}
