% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentDownload.R
\name{edgar_prepare_print_message}
\alias{edgar_prepare_print_message}
\title{Prepare Print Messages for a Nested EDGAR Download Group}
\usage{
edgar_prepare_print_message(.nst, .counter)
}
\arguments{
\item{.nst}{A tibble (usually the output of \code{tidyr::nest()}) with at least the following columns:}
}
\value{
A tibble
}
\description{
Prepare Print Messages for a Nested EDGAR Download Group
}
\keyword{internal}
