% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{filter_edgar_data}
\alias{filter_edgar_data}
\title{Filter EDGAR Data by CIK and Document Type}
\usage{
filter_edgar_data(.tab, .params)
}
\arguments{
\item{.tab}{Data frame or tibble containing EDGAR data}

\item{.params}{List of parameters from get_edgar_params()}
}
\value{
Filtered data frame or tibble
}
\description{
Internal helper function to filter EDGAR data based on CIK numbers and document types
}
\keyword{internal}
