% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{format_number}
\alias{format_number}
\title{Format Numbers for Display}
\usage{
format_number(.num, .type = c("c", "p"))
}
\arguments{
\item{.num}{Numeric value to format}

\item{.type}{Character string specifying format type: "c" for comma or "p" for percentage}
}
\value{
Formatted character string
}
\description{
Formats numbers as either comma-separated values or percentages.
}
\keyword{internal}
