% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{format_time}
\alias{format_time}
\title{Format Time Status Message}
\usage{
format_time(.t0, .i, .n)
}
\arguments{
\item{.t0}{POSIXct start time}

\item{.i}{Current iteration number}

\item{.n}{Total number of iterations}
}
\value{
Formatted status message
}
\description{
Creates a formatted status message showing elapsed time and estimated time remaining.
}
\keyword{internal}
