% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FormTypesMod}
\alias{FormTypesMod}
\title{SEC EDGAR Form Types Dataset (Modified)}
\format{
A tibble with 4 variables:
\describe{
\item{FormClass}{character. Classification category of the form}
\item{FormTypeMod}{character. Standardized version of the form type}
\item{nDocs}{integer. Aggregated count of occurrences}
\item{Description}{character. Detailed description of the form type's purpose}
}
}
\source{
SEC EDGAR database master index files
}
\usage{
FormTypesMod
}
\description{
A summarized dataset of SEC EDGAR form types grouped by their classifications
and standardized versions.
}
\keyword{EDGAR}
\keyword{SEC}
\keyword{datasets}
\keyword{filings}
\keyword{forms}
\keyword{standardized}
