% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_header}
\alias{get_header}
\title{Get HTTP Request Header}
\usage{
get_header(.user)
}
\arguments{
\item{.user}{User agent string for HTTP requests}
}
\value{
httr header object
}
\description{
Creates HTTP request headers for SEC EDGAR API calls.
}
\keyword{internal}
