% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentLinks.R
\name{get_landing_html}
\alias{get_landing_html}
\title{Get HTML Content of SEC EDGAR Filing Index Page}
\usage{
get_landing_html(.url, .user, .verbose, .path_log)
}
\arguments{
\item{.url}{Character string specifying the URL of the filing index page to download.}

\item{.user}{Character string specifying the user agent to be used in HTTP requests.
This should typically be an email address to comply with SEC's fair access policy.}

\item{.verbose}{Logical indicating whether to print progress messages to the console.
Default is FALSE.}
}
\value{
A tibble with a single column 'HTML' containing the HTML content of the requested page.
If an error occurs, returns an empty tibble with NA value.
}
\description{
Downloads the HTML content of an SEC EDGAR filing index page. This function
handles the HTTP request and error checking for retrieving filing index pages.
}
\details{
The function makes an HTTP GET request to the specified URL using the provided
user agent. It includes error handling to manage network issues or invalid responses.

If the request is successful, the function returns the HTML content as a tibble
with a single row and column. If an error occurs, the function returns an empty
tibble with an NA value, allowing the calling function to continue processing
other URLs without interruption.
}
\keyword{internal}
