% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_processed_documents}
\alias{get_processed_documents}
\title{Get Processed Documents}
\usage{
get_processed_documents(.dir)
}
\arguments{
\item{.dir}{Directory containing processed documents}
}
\value{
Vector of processed document hashes
}
\description{
Retrieves list of documents that have already been processed.
}
\keyword{internal}
