% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_processed_hash_index}
\alias{get_processed_hash_index}
\title{Get Processed Hash Indexes}
\usage{
get_processed_hash_index(.con)
}
\arguments{
\item{.con}{Database connection object}
}
\value{
Vector of processed hash indexes
}
\description{
Retrieves list of hash indexes that have already been processed.
}
\keyword{internal}
