% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_processed_year_qtr}
\alias{get_processed_year_qtr}
\title{Get Processed Year-Quarters}
\usage{
get_processed_year_qtr(.con)
}
\arguments{
\item{.con}{Database connection object}
}
\value{
Vector of processed year-quarters
}
\description{
Retrieves list of year-quarters that have already been processed.
}
\keyword{internal}
