% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentDownload.R
\name{get_tbp_docids}
\alias{get_tbp_docids}
\title{Prepare "To Be Processed" Document IDs Queue}
\usage{
get_tbp_docids(.dir, .doc_ids, .workers)
}
\arguments{
\item{.dir}{Character string specifying the root directory where SEC EDGAR data is stored.}

\item{.doc_ids}{Character vector of document IDs to process. If NULL, all available documents
that haven't been previously downloaded will be processed.}

\item{.workers}{Integer specifying the number of parallel workers for downloading.}
}
\value{
A tibble with distinct combinations of DocType, YearQuarter, Seq, dir_tmp,
and dir_out, representing the batches to be processed.
}
\description{
Prepares a queuing system for SEC EDGAR documents to be downloaded in batches.
This function creates a temporary SQLite database to track which documents
need to be processed and assigns them to download batches.
}
\keyword{internal}
