% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_to_be_processed_download_links}
\alias{get_to_be_processed_download_links}
\title{Get Document Links To Be Downloaded}
\usage{
get_to_be_processed_download_links(.dir, .params, .workers)
}
\arguments{
\item{.dir}{Base directory for the SEC data storage}

\item{.params}{List of filtering parameters from get_edgar_params()}

\item{.workers}{Number of parallel workers to use for processing}
}
\value{
Writes a parquet file to the temporary directory containing links to be processed,
with an additional 'Split' column for parallel processing. Only includes
documents with valid file extensions.
}
\description{
Identifies and prepares document links that haven't been downloaded yet. The function:
\enumerate{
\item Checks already downloaded documents in the main document data directory
\item Filters the document links dataset based on provided parameters
\item Excludes already downloaded documents
\item Prepares the data for parallel processing
}
}
\keyword{internal}
