% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_warning_document_download}
\alias{get_warning_document_download}
\title{Generate Document Download Warning}
\usage{
get_warning_document_download(.cik)
}
\arguments{
\item{.cik}{CIK number of the failed download}
}
\description{
Generates a warning message for document download failures.
}
\keyword{internal}
