% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_warning_document_links}
\alias{get_warning_document_links}
\title{Generate Document Links Warning}
\usage{
get_warning_document_links(.cik)
}
\arguments{
\item{.cik}{CIK number of the failed parsing}
}
\description{
Generates a warning message for document links parsing failures.
}
\keyword{internal}
