% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{get_warning_master_index}
\alias{get_warning_master_index}
\title{Generate Master Index Warning}
\usage{
get_warning_master_index(.year, .qtr)
}
\arguments{
\item{.year}{Year of the failed fetch}

\item{.qtr}{Quarter of the failed fetch}
}
\description{
Generates a warning message for master index fetch failures.
}
\keyword{internal}
