% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{help_download_document}
\alias{help_download_document}
\title{Download and Save SEC EDGAR Documents}
\usage{
help_download_document(.doc_row, .user)
}
\arguments{
\item{.doc_row}{A row with download information}

\item{.user}{Character string of the user agent}
}
\value{
A Document
}
\description{
Downloads and saves SEC EDGAR documents in appropriate formats based on their file extension.
Handles text-based documents (htm, xml, xsd, txt) by converting them to parquet,
and binary files (pdf, gif, jpg) by saving them directly.
}
\keyword{internal}
