% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-External.R
\name{html_options}
\alias{html_options}
\title{Create HTML reading options}
\usage{
html_options(
  .recover = TRUE,
  .huge = TRUE,
  .noerror = FALSE,
  .nowarning = FALSE,
  .encoding = "UTF-8"
)
}
\arguments{
\item{.recover}{Logical; should parser try to recover from errors? (default: TRUE)}

\item{.huge}{Logical; optimize for huge files? (default: TRUE)}

\item{.noerror}{Logical; suppress error messages? (default: FALSE)}

\item{.nowarning}{Logical; suppress warning messages? (default: FALSE)}

\item{.encoding}{Character encoding to use (default: "UTF-8")}
}
\value{
List of HTML reading options
}
\description{
Create HTML reading options
}
