% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{initial_landing_page_database}
\alias{initial_landing_page_database}
\title{Initialize Document Links Database}
\usage{
initial_landing_page_database(.dir)
}
\arguments{
\item{.dir}{Path to SQLite database file}
}
\description{
Creates SQLite database schema for document links if it doesn't exist.
}
\keyword{internal}
