% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{initial_master_index_database}
\alias{initial_master_index_database}
\title{Initialize Master Index Database}
\usage{
initial_master_index_database(.path)
}
\arguments{
\item{.path}{Path to SQLite database file}
}
\description{
Creates SQLite database schema for master index if it doesn't exist.
}
\keyword{internal}
