% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{make_get_request}
\alias{make_get_request}
\title{Make HTTP GET Request}
\usage{
make_get_request(.url, .user)
}
\arguments{
\item{.url}{URL to request}

\item{.user}{User agent string}
}
\value{
httr response object or error
}
\description{
Makes an HTTP GET request with error handling.
}
\keyword{internal}
