% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModTypes}
\alias{ModTypes}
\title{SEC EDGAR Modified Form-Document Type Combinations Dataset}
\format{
A tibble with 3 variables:
\describe{
\item{FormTypeMod}{character. The modified/standardized form type}
\item{DocTypeMod}{character. The modified/standardized document type}
\item{nModTypes}{integer. Count of occurrences of this combination}
}
}
\source{
SEC EDGAR database master index and document indices
}
\usage{
ModTypes
}
\description{
A dataset containing all observed combinations of standardized form types and
document types from SEC EDGAR filings. This dataset is similar to RawTypes but
uses the modified/standardized versions of both form and document types.
}
\details{
This dataset is created by joining the master index with document links using
HashIndex as the key, using the modified versions of both form and document types.
Currently, the modified versions are identical to the raw versions, serving as
placeholders for future standardization efforts.
}
\keyword{EDGAR}
\keyword{SEC}
\keyword{combinations}
\keyword{datasets}
\keyword{documents}
\keyword{filings}
\keyword{standardized}
