% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-External.R
\name{read_html}
\alias{read_html}
\title{Read and extract text from HTML content}
\usage{
read_html(.html, .options = html_options())
}
\arguments{
\item{.html}{Character string containing HTML content or path to HTML file}

\item{.options}{List of options from html_options()}
}
\value{
Character string containing extracted text; NA_character_ if parsing fails
}
\description{
Read and extract text from HTML content
}
