% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-External.R
\name{standardize_string}
\alias{standardize_string}
\title{Standardize String Content}
\usage{
standardize_string(.str)
}
\arguments{
\item{.str}{Character string to standardize}
}
\value{
Standardized character string
}
\description{
Standardizes string content by removing extra whitespace and converting to ASCII.
}
\keyword{internal}
