% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table_DocTypesMod}
\alias{Table_DocTypesMod}
\title{SEC EDGAR Document Types Dataset (Modified)}
\format{
A tibble with 4 variables:
\describe{
\item{DocClass}{character. Classification category of the document}
\item{DocTypeMod}{character. Standardized version of the document type}
\item{nDocs}{integer. Aggregated count of occurrences}
\item{Description}{character. Detailed description of the document type's purpose}
}
}
\source{
SEC EDGAR database document indices
}
\usage{
Table_DocTypesMod
}
\description{
A summarized dataset of SEC EDGAR document types grouped by their classifications
and standardized versions.
}
\keyword{EDGAR}
\keyword{SEC}
\keyword{datasets}
\keyword{documents}
\keyword{exhibits}
\keyword{filings}
\keyword{standardized}
