% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{write_landing_page_to_parquet}
\alias{write_landing_page_to_parquet}
\title{Write Document Links to Parquet File}
\usage{
write_landing_page_to_parquet(.dir)
}
\arguments{
\item{.dir}{Base directory for the SEC data storage}
}
\description{
Transfers the entire document links table from SQLite database to a Parquet file format.
This function provides a way to maintain a Parquet copy of the document links data,
which can be more efficient for certain types of data analysis and filtering operations.
}
\keyword{internal}
