% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{write_landing_page_to_sqlite}
\alias{write_landing_page_to_sqlite}
\title{Write Document Links to SQLite Database}
\usage{
write_landing_page_to_sqlite(.dir, .tab)
}
\arguments{
\item{.dir}{Base directory for the SEC data storage}

\item{.tab}{Data frame or tibble containing new document links to be appended.
Must match the structure of the 'document_links' table.}
}
\description{
Appends new document links data to the existing SQLite database. This function
serves as a way to incrementally add new document links to the persistent storage.
}
\keyword{internal}
