% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{content_to_dataframe}
\alias{content_to_dataframe}
\title{Convert Content to DataFrame}
\usage{
content_to_dataframe(.content, .year, .qtr)
}
\arguments{
\item{.content}{Raw content string}

\item{.year}{Year of the content}

\item{.qtr}{Quarter of the content}
}
\value{
Tibble with processed master index data
}
\description{
Converts raw master index content to a structured data frame.
}
\keyword{internal}
