% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{help_download_document}
\alias{help_download_document}
\title{Download Document}
\usage{
help_download_document(.url, .user)
}
\arguments{
\item{.url}{URL of the document}

\item{.user}{User agent string}
}
\value{
Tibble with document content and processed text
}
\description{
Downloads and processes a single document.
}
\keyword{internal}
