% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers-Internal.R
\name{help_get_document_link}
\alias{help_get_document_link}
\title{Get Document Link}
\usage{
help_get_document_link(.url, .user)
}
\arguments{
\item{.url}{URL of the index page}

\item{.user}{User agent string}
}
\value{
Tibble with document links data
}
\description{
Downloads and processes document links from an index page.
}
\keyword{internal}
